realtime user table

+++ app/Livewire/Modals/User/Delete.php
@@ -4,6 +4,7 @@ namespace App\Livewire\Modals\User;
 
 use App\Http\Controllers\UserController;
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Attributes\On;
 use Livewire\Component;
 
@@ -26,12 +27,23 @@ class Delete extends Component
 
     public function delete() {
         $this->userController = new UserController();
-        $user = $this->userController->destroy($this->student->id);
 
-        if ($user) {
-            return redirect()->route('management-students');
+        if($this->student->id != Auth::id()) {
+            $user = $this->userController->destroy($this->student->id);
+
+            if ($user) {
+                session()->flash('error', [
+                    'title' => 'Berhasil',
+                    'message' => 'User Telah Dihapus'
+                ]);
+                return $this->redirect(request()->header('Referer'));
+            }
+        } else {
+            return session()->flash('error', [
+                'title' => 'Gagal',
+                'message' => 'Tidak dapat menghapus diri sendiri.'
+            ]);
         }
-
     }
 
     public function render()
